Update ansible to install correct version of yarn
This script installs yarn 0.18 by the time it's written but it will install version 0.23 as of now -- which will break if running on top of node 4.x. (Yarn 0.23 is recommended to run on node 6.x.)

This edit makes sure that yarn version 0.18 is installed.

TODO: node 4.x now becomes 4.8 instead of 4.3.3 as of the time the script is written. This may potentially break something if we setup a new server.