[f] 014 - Fix disappearing body when modal is active
In order to prevent vertical scrolling when generic modal is active, we
added `overflow: hidden` style to the main content’s wrapper. Problems
with this approach are:

- Empty page content when modal is active: because the wrapper's
  natural height is 0px (due to page transition wrapper being `absolute`
  positioned), setting `overflow: hidden` will hide all of its inner
  content.

- The page is always scrolled to top when modal is open.

- The page content is shifted to the right when modal is activated
  because the vertical scrollbar disappeared.

Applying `overflow: hidden` directly on <body> solves the first 2
problems. For the last one, we manually add padding-right to it to
make up for the scrollbar's width.

[#151286014] https://www.pivotaltracker.com/story/show/151286014